modularise code
